Refactor Map into its own component
